really store the event
