+++ cmd/metal-api/internal/grpc/wait-service_int_test.go
@@ -101,6 +101,9 @@ func (ds *datasource) UpdateMachine(old, new *metal.Machine) error {
 func (ds *datasource) CreateMachine(new *metal.Machine) error {
 	return nil
 }
+func (ds *datasource) ProvisioningEventForMachine(machineID, event, message string) (*metal.ProvisioningEventContainer, error) {
+	return nil, nil
+}
 
 func (t *test) run() {
 	defer t.shutdown()
